feat: branding and layout changes, landing page routing fixes.
