+++ src/components/ui/footer.tsx
@@ -1,15 +1,38 @@
 import { Link } from "react-router-dom";
 
 export const Footer = () => {
+  const handleFeatureClick = (e: React.MouseEvent<HTMLAnchorElement>) => {
+    e.preventDefault();
+    const featuresSection = document.getElementById('features');
+    if (featuresSection) {
+      featuresSection.scrollIntoView({ behavior: 'smooth' });
+    }
+  };
+
+  const handlePricingClick = (e: React.MouseEvent<HTMLAnchorElement>) => {
+    e.preventDefault();
+    const featuresSection = document.getElementById('pricing');
+    if (featuresSection) {
+      featuresSection.scrollIntoView({ behavior: 'smooth' });
+    }
+  };
+  
+  const handleFaqClick = (e: React.MouseEvent<HTMLAnchorElement>) => {
+    e.preventDefault();
+    const featuresSection = document.getElementById('faq');
+    if (featuresSection) {
+      featuresSection.scrollIntoView({ behavior: 'smooth' });
+    }
+  };
+
+
   return (
     <footer className="border-t py-8">
       <div className="container px-4">
         <div className="grid gap-8 md:grid-cols-4">
           <div className="space-y-4">
             <Link to="/" className="flex items-center gap-2">
-              <div className="w-8 h-8 rounded-full bg-gradient-to-br from-primary to-secondary flex items-center justify-center text-white font-bold">
-                Z
-              </div>
+            <img src="noma-logo.svg" alt="Zencora Noma Logo" className="w-10 h-10" />
               <span className="font-bold text-xl">Zencora Noma</span>
             </Link>
             <p className="text-sm text-muted-foreground">
@@ -21,13 +44,18 @@ export const Footer = () => {
             <h4 className="font-semibold mb-4">Produto</h4>
             <ul className="space-y-2 text-sm">
               <li>
-                <Link to="/features" className="text-muted-foreground hover:text-foreground transition-colors">
+                <Link to="/#features" onClick={handleFeatureClick} className="text-muted-foreground hover:text-foreground transition-colors">
                   Funcionalidades
                 </Link>
               </li>
               <li>
-                <Link to="/pricing" className="text-muted-foreground hover:text-foreground transition-colors">
+                <Link to="/pricing" onClick={handlePricingClick} className="text-muted-foreground hover:text-foreground transition-colors">
                   Preços
+                </Link>
+              </li>
+              <li>
+                <Link to="/faq" onClick={handleFaqClick} className="text-muted-foreground hover:text-foreground transition-colors">
+                  FAQ
                 </Link>
               </li>
             </ul>
